extract SectionDivider widget
